refactor: modularize bot and add containerization
