+++ README.md
@@ -10,6 +10,12 @@
 
 ## Запуск
 
+### Docker
+
+1. Скопируйте `.env.example` в `.env` и задайте значение `TELEGRAM_BOT_TOKEN`.
+2. Запустите бота командой `docker-compose up --build`.
+
+
 1. Склонируйте этот репозиторий на свой локальный компьютер.
 2. Скачайте базы данных IP адресов `GeoLite2` (файлы называется `GeoLite2-City.mmdb` и `GeoLite2-ASN.mmdb`) в папку с репозиторием.
 3. Установите необходимые пакеты Python (см. импорты в файле бота)
