+++ docker-compose.yml
@@ -1,0 +1,9 @@
+version: '3.8'
+services:
+  bot:
+    build: .
+    env_file:
+      - .env
+    volumes:
+      - ./:/app
+    restart: unless-stopped
