+++ requirements.txt
@@ -1,0 +1,4 @@
+python-telegram-bot==13.15
+requests
+geoip2
+openpyxl
